Animation init commit

+++ Snake/AnalogClock/consts.h
@@ -2,11 +2,11 @@
 
 static const unsigned SCREEN_WIDTH = 1000;
 static const unsigned SCREEN_HEIGHT = 600;
-static const unsigned CELL_SIZE = 10;
+static const float CELL_SIZE = 10;
 static const unsigned INITIAL_NUMBER_OF_BODY_PARTS = 10;
 static const unsigned SNAKE_USIAL_INCREASE = 5;
 static const unsigned SNAKE_DECREASE = 4;
 static const unsigned MESSAGE_WINDOW_HEIGHT = 100;
 static const unsigned MESSAGE_WINDOW_WIDTH = 300;
 static const unsigned MESSAGE_FONT_SIZE = 23;
-static const float ITER_TIME = 0.1;
+static const float ITER_TIME = 0.1f;
